pyproject.toml
[project]
name = "pdf-analyzer"
version = "0.2.0"
description = "Análisis híbrido de errores en PDFs usando LanguageTool + Ollama LLM"
requires-python = ">=3.12"
dependencies = [
    "language-tool-python>=2.9.4",
    "ollama>=0.6.0",
    "pdfminer-six>=20250506",
    "tqdm>=4.67.1",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
